# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ranks": {
      "get": {
        "operationId": "RanksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "f1a7c3e9-5b2d-4806-9e4f-a3c1d7b5e290",
                    "name": "Wood",
                    "minXp": 0,
                    "badgeUrl": "/uploads/images/9b3e5f71-2c8d-4a06-b4e9-5d1a7c0f2e68.webp",
                    "createdAt": "2026-05-20T09:30:00.000Z",
                    "updatedAt": "2026-05-20T09:30:00.000Z"
                  },
                  {
                    "id": "f1a7c3e9-5b2d-4806-9e4f-a3c1d7b5e290",
                    "name": "Gold",
                    "minXp": 4000,
                    "badgeUrl": "/uploads/images/9b3e5f71-2c8d-4a06-b4e9-5d1a7c0f2e68.webp",
                    "createdAt": "2026-05-20T09:30:00.000Z",
                    "updatedAt": "2026-05-20T09:30:00.000Z"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          }
        },
        "tags": [
          "Ranks"
        ],
        "summary": "List (ranks)",
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "Needs a bearer token. Minimum scope: **self** (any signed-in user)."
      },
      "post": {
        "operationId": "RanksController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": "f1a7c3e9-5b2d-4806-9e4f-a3c1d7b5e290",
                  "name": "Gold",
                  "minXp": 4000,
                  "badgeUrl": "/uploads/images/9b3e5f71-2c8d-4a06-b4e9-5d1a7c0f2e68.webp",
                  "createdAt": "2026-05-20T09:30:00.000Z",
                  "updatedAt": "2026-05-20T09:30:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "request failed validation",
                  "code": "VALIDATION_ERROR",
                  "issues": [
                    {
                      "path": "machineId",
                      "message": "Invalid UUID"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          },
          "403": {
            "description": "Caller scope is too low for this endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Forbidden",
                  "code": "FORBIDDEN"
                }
              }
            }
          }
        },
        "tags": [
          "Ranks"
        ],
        "summary": "Create (ranks)",
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "Needs a bearer token. Minimum scope: **admin** (admin).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "minXp": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100000000
                  },
                  "badgeUrl": {
                    "nullable": true,
                    "type": "string",
                    "pattern": "^\\/uploads\\/images\\/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\\.webp$"
                  }
                },
                "required": [
                  "name",
                  "minXp"
                ]
              },
              "example": {
                "name": "string",
                "minXp": 1
              }
            }
          }
        }
      }
    },
    "/ranks/{id}": {
      "patch": {
        "operationId": "RanksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "3f2b8c1e-7a4d-4e5b-9c61-0d2f8a9b1c34"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": "f1a7c3e9-5b2d-4806-9e4f-a3c1d7b5e290",
                  "name": "Gold",
                  "minXp": 4000,
                  "badgeUrl": "/uploads/images/9b3e5f71-2c8d-4a06-b4e9-5d1a7c0f2e68.webp",
                  "createdAt": "2026-05-20T09:30:00.000Z",
                  "updatedAt": "2026-05-20T09:30:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "request failed validation",
                  "code": "VALIDATION_ERROR",
                  "issues": [
                    {
                      "path": "machineId",
                      "message": "Invalid UUID"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          },
          "403": {
            "description": "Caller scope is too low for this endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Forbidden",
                  "code": "FORBIDDEN"
                }
              }
            }
          }
        },
        "tags": [
          "Ranks"
        ],
        "summary": "Update (ranks)",
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "Needs a bearer token. Minimum scope: **admin** (admin).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "minXp": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100000000
                  },
                  "badgeUrl": {
                    "nullable": true,
                    "type": "string",
                    "pattern": "^\\/uploads\\/images\\/[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\\.webp$"
                  }
                }
              },
              "example": {}
            }
          }
        }
      },
      "delete": {
        "operationId": "RanksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "3f2b8c1e-7a4d-4e5b-9c61-0d2f8a9b1c34"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": "f1a7c3e9-5b2d-4806-9e4f-a3c1d7b5e290",
                  "deleted": true
                }
              }
            }
          },
          "400": {
            "description": "Request failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "request failed validation",
                  "code": "VALIDATION_ERROR",
                  "issues": [
                    {
                      "path": "machineId",
                      "message": "Invalid UUID"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "code": "UNAUTHORIZED"
                }
              }
            }
          },
          "403": {
            "description": "Caller scope is too low for this endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Forbidden",
                  "code": "FORBIDDEN"
                }
              }
            }
          }
        },
        "tags": [
          "Ranks"
        ],
        "summary": "Remove (ranks)",
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "Needs a bearer token. Minimum scope: **admin** (admin)."
      }
    },
...
  },
  "info": {
    "title": "cstam backend",
    "description": "REST API of the gaming-cafe platform: accounts, bookings, PIN check-in, sessions, billing and wallet.\n\n**Flow to try**\n1. `POST /users` create a gamer (or use a seeded account), then `POST /auth/login`.\n2. Copy `accessToken`, click **Authorize**, paste it (no \"Bearer \" prefix).\n3. `GET /machines` or `GET /branches/{id}/stations` pick a machine.\n4. `POST /reservations` book it, the answer holds the station PIN.\n5. Staff top up with `POST /wallets/{gamerProfileId}/credit`; sessions are billed from the wallet.\n\nErrors always have the shape `{ error, code, issues? }`. Money is an integer in the smallest unit.",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "request failed validation"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}